+++ packages/gatsby-cli/src/structured-errors/error-map.js
@@ -89,6 +89,77 @@ const errorMap = {
     type: `PLUGIN`,
     level: `ERROR`,
   },
+  "11322": {
+    text: context =>
+      `${
+        context.pluginName
+      } created a page and didn't pass the path to the component.\n\nThe page object passed to createPage:\n${JSON.stringify(
+        context.pageObject,
+        null,
+        4
+      )}\n\nSee the documentation for the "createPage" action — https://www.gatsbyjs.org/docs/actions/#createPage`,
+    level: `ERROR`,
+  },
+  "11323": {
+    text: context =>
+      `${
+        context.pluginName
+      } must set the page path when creating a page.\n\nThe page object passed to createPage:\n${JSON.stringify(
+        context.pageObject,
+        null,
+        4
+      )}\n\nSee the documentation for the "createPage" action — https://www.gatsbyjs.org/docs/actions/#createPage`,
+    level: `ERROR`,
+  },
+  "11324": {
+    text: context =>
+      `${
+        context.message
+      }\n\nSee the documentation for the "createPage" action — https://www.gatsbyjs.org/docs/actions/#createPage`,
+    level: `ERROR`,
+  },
+  "11325": {
+    text: context =>
+      `${
+        context.pluginName
+      } created a page with a component that doesn't exist.\n\nThe path to the missing component is "${
+        context.component
+      }"\n\nThe page object passed to createPage:\n${JSON.stringify(
+        context.pageObject,
+        null,
+        4
+      )}\n\nSee the documentation for the "createPage" action — https://www.gatsbyjs.org/docs/actions/#createPage`,
+    level: `ERROR`,
+  },
+  "11326": {
+    text: context =>
+      `${
+        context.pluginName
+      } must set the absolute path to the page component when create creating a page.\n\nThe (relative) path you used for the component is "${
+        context.component
+      }"\n\nYou can convert a relative path to an absolute path by requiring the path module and calling path.resolve() e.g.\n\nconst path = require("path")\npath.resolve("${
+        context.component
+      }")\n\nThe page object passed to createPage:\n${JSON.stringify(
+        context.pageObject,
+        null,
+        4
+      )}\n\nSee the documentation for the "createPage" action — https://www.gatsbyjs.org/docs/actions/#createPage`,
+    level: `ERROR`,
+  },
+  "11327": {
+    text: context =>
+      `You have an empty file in the "src/pages" directory at "${
+        context.relativePath
+      }". Please remove it or make it a valid component`,
+    level: `ERROR`,
+  },
+  "11328": {
+    text: context =>
+      `A page component must export a React component for it to be valid. Please make sure this file exports a React component:\n\n${
+        context.fileName
+      }`,
+    level: `ERROR`,
+  },
 }
 
 module.exports = { errorMap, defaultError: errorMap[``] }